(feat) show mvp mobile navigation
