+++ src/layout/Navigation/MobileNavigation.js
@@ -1,0 +1,12 @@
+import React from "react";
+// import { Drawer, List, ListItem } from "@mui/material";
+
+const MobileNavigation = () => {
+  return (
+    <>
+      <p>Mobile Nav</p>
+    </>
+  );
+};
+
+export default MobileNavigation;
