+++ src/layout/Navigation/Navigation.js
@@ -1,35 +1,23 @@
 import React from "react";
-import { useRouter } from "next/router";
-import { Grid } from "@mui/material";
+import { useTheme } from "@mui/material/styles";
+
+import { useMediaQuery } from "@mui/material";
 import { navLinks } from "gertrudenyenyeshi/config";
-import { StyledLink, NavBar } from "./NavigationStyles";
+import { NavBar } from "./NavigationStyles";
+import MobileNavigation from "./MobileNavigation";
+import DesktopNavigation from "./DesktopNavigation";
 
 function Navigation() {
-  const router = useRouter();
-  const currentRoute = router.pathname;
+  const theme = useTheme();
+  const lgScreen = useMediaQuery(theme.breakpoints.up("md"));
 
   return (
     <NavBar>
-      <Grid
-        container
-        justifyContent="center"
-        alignItems="center"
-        direction={{ xs: "column", sm: "row" }}
-        sx={{
-          height: "100%",
-          padding: { xs: "20px 0", md: 0 },
-        }}
-      >
-        {navLinks.map(({ href, name }) => (
-          <StyledLink
-            className={currentRoute === href ? "active" : ""}
-            key={href}
-            href={href}
-          >
-            {name}
-          </StyledLink>
-        ))}
-      </Grid>
+      {lgScreen ? (
+        <DesktopNavigation navLinks={navLinks} />
+      ) : (
+        <MobileNavigation navLinks={navLinks} />
+      )}
     </NavBar>
   );
 }
